feat: reorganized the api endpoints
